Correction de bugs multiples sur l'interface de gestion des statuts des intervenants

+++ module/Application/src/Application/Controller/StatutIntervenantController.php
@@ -4,17 +4,19 @@ namespace Application\Controller;
 
 use Application\Entity\Db\StatutIntervenant;
 use Application\Provider\Privilege\Privileges;
-use Application\Service\Traits\StatutIntervenantAwareTrait;
 use Application\Exception\DbException;
 use Application\Form\StatutIntervenant\Traits\StatutIntervenantSaisieFormAwareTrait;
+use Application\Service\Traits\StatutIntervenantServiceAwareTrait;
 use UnicaenApp\View\Model\MessengerViewModel;
 use Application\Service\Traits\TypeIntervenantServiceAwareTrait;
 
 class StatutIntervenantController extends AbstractController
 {
-    use StatutIntervenantAwareTrait;
+    use StatutIntervenantServiceAwareTrait;
     use StatutIntervenantSaisieFormAwareTrait;
     use TypeIntervenantServiceAwareTrait;
+
+
 
     public function indexAction()
     {
@@ -28,27 +30,27 @@ class StatutIntervenantController extends AbstractController
     }
 
 
+
     public function saisieAction()
     {
         /* @var $statutIntervenant StatutIntervenant */
 
         $statutIntervenant = $this->getEvent()->getParam('statutIntervenant');
-        $form = $this->getFormStatutIntervenantSaisie();
+        $form              = $this->getFormStatutIntervenantSaisie();
         if (empty($statutIntervenant)) {
-            //$title = 'Création d\'un nouveau statut d\'intervenant';
-            $tblSi=$this->getServiceStatutIntervenant()->getList();
-            $ordre=0;
-            foreach($tblSi as $eltSi){
-                if ($eltSi->getOrdre() > $ordre) $ordre=$eltSi->getOrdre();
-            }
-            $ordre++;
+            $title             = 'Création d\'un nouveau statut d\'intervenant';
             $statutIntervenant = $this->getServiceStatutIntervenant()->newEntity();
-            $statutIntervenant->setOrdre($this->getServiceStatutIntervenant()->fetchMaxOrdre()+1);
+            $statutIntervenant->setOrdre($this->getServiceStatutIntervenant()->fetchMaxOrdre() + 1);
         } else {
             $title = 'Édition d\'un statut d\'intervenant';
         }
 
-        if ($this->isAllowed(Privileges::getResourceId(Privileges::INTERVENANT_STATUT_EDITION))) {
+        $canEdit = $this->isAllowed(Privileges::getResourceId(Privileges::INTERVENANT_STATUT_EDITION));
+        if ($statutIntervenant->getSource()->getImportable()) {
+            $canEdit = false; // Si la source est synchronisable alors pas d'édition possible
+        }
+
+        if ($canEdit) {
             $form->bindRequestSave($statutIntervenant, $this->getRequest(), function (StatutIntervenant $si) {
                 try {
                     $this->getServiceStatutIntervenant()->save($si);
@@ -58,7 +60,7 @@ class StatutIntervenantController extends AbstractController
                     $this->flashMessenger()->addErrorMessage($e->getMessage() . ':' . $si->getId());
                 }
             });
-        }else{
+        } else {
             $form->bind($statutIntervenant);
             $form->readOnly();
         }
@@ -67,68 +69,73 @@ class StatutIntervenantController extends AbstractController
         return compact('form', 'title');
     }
 
+
+
     public function deleteAction()
     {
         $statutIntervenant = $this->getEvent()->getParam('statutIntervenant');
 
-        try {
-            $this->getServiceStatutIntervenant()->delete($statutIntervenant);
-            $this->flashMessenger()->addSuccessMessage("Statut d'Intervenant supprimé avec succès.");
-        } catch (\Exception $e) {
-            $this->flashMessenger()->addErrorMessage(DbException::translate($e)->getMessage());
+        $canEdit = $this->isAllowed(Privileges::getResourceId(Privileges::INTERVENANT_STATUT_EDITION));
+        if ($statutIntervenant->getSource()->getImportable()) {
+            $canEdit = false; // Si la source est synchronisable alors pas d'édition possible
         }
+
+        if (!$canEdit) {
+            $this->flashMessenger()->addErrorMessage('Statut non modifiable : droit non accordé, car vous n\'avez pas le privilège pour cela ou bien le statut est synchronisé depuis un autre logiciel');
+        } else {
+            try {
+                $this->getServiceStatutIntervenant()->delete($statutIntervenant);
+                $this->flashMessenger()->addSuccessMessage("Statut d'Intervenant supprimé avec succès.");
+            } catch (\Exception $e) {
+                $this->flashMessenger()->addErrorMessage(DbException::translate($e)->getMessage());
+            }
+        }
+
         return new MessengerViewModel(compact('statutIntervenant'));
     }
 
+
+
     public function statutIntervenantTrierAction()
     {
-        /* @var $si StatutIntervenant */
-        $txt='result=';
         $champsIds = explode(',', $this->params()->fromPost('champsIds', ''));
-        $this->flashMessenger(var_dump($champsIds));
-        $ordre = 1;
+        $ordre     = $this->getServiceStatutIntervenant()->fetchMaxOrdre() + 1; // Pour éviter tout doublon!!
         foreach ($champsIds as $champId) {
-            $txt.=$champId.'=>';
             $si = $this->getServiceStatutIntervenant()->get($champId);
             if ($si) {
-                $txt .= ';' . $si->getOrdre();
                 $si->setOrdre($ordre);
                 $ordre++;
                 try {
                     $this->getServiceStatutIntervenant()->save($si);
                 } catch (\Exception $e) {
-                    $e = DbException::translate($e);
-                    $txt .= ':' . $e->getMessage();
+                    $this->flashMessenger()->addErrorMessage(DbException::translate($e)->getMessage());
                 }
             }
         }
+
         return new JsonModel(['msg' => 'Tri des champs effectué']);
     }
+
+
 
     public function cloneAction()
     {
         /* @var $statutIntervenant StatutIntervenant */
-        /* @var $statutIntervenantNew StatutIntervenant */
-
         $statutIntervenant = $this->getEvent()->getParam('statutIntervenant');
-        $form = $this->getFormStatutIntervenantSaisie();
-        if (empty($statutIntervenant)) {
-            $this->flashMessenger()->addErrorMessage($e->getMessage() . ':' . $statutIntervenant->getId());
-            return;
-        }
-        $title = 'Dupplication d\'un statut d\'intervenant';
-        $statutIntervenantNew = clone $statutIntervenant;
-        $i=$this->getServiceStatutIntervenant()->fetchMaxOrdre()+1;
-        $statutIntervenantNew->setOrdre($i+1);
+        $form              = $this->getFormStatutIntervenantSaisie();
+
+        $title                = 'Duplication d\'un statut d\'intervenant';
+        $newStatutIntervenant = $statutIntervenant->dupliquer();
+        $newStatutIntervenant->setOrdre($this->getServiceStatutIntervenant()->fetchMaxOrdre() + 1);
         //$statutIntervenantNew->setOrdre($this->getServiceStatutIntervenant()->fetchMaxOrdre()+1);
-        $form->bind($statutIntervenantNew);
+        $form->bind($newStatutIntervenant);
         $request = $this->getRequest();
         if ($request->isPost()) {
             $form->setData($request->getPost());
             if ($form->isValid()) {
                 try {
-                    $this->getServiceStatutIntervenant()->save($statutIntervenantNew);
-                    $form->get('id')->setValue($statutIntervenantNew->getId()); // transmet le nouvel ID
+                    $this->getServiceStatutIntervenant()->save($newStatutIntervenant);
+                    $form->get('id')->setValue($newStatutIntervenant->getId()); // transmet le nouvel ID
                     $this->flashMessenger()->addSuccessMessage('Enregistrement effectué');
                 } catch (\Exception $e) {
                     $e        = DbException::translate($e);
